Icons: Refactor 14x13 Icon Rendering to use one font
Icon drawing now uses the same 14x13 compacticons font file.
This change also brings icons to more menus within Themify and updates most of the home menu icons.

The menu item for loading the WPS on the home screen is now also adaptive.
It will display a play icon when 'Resume Playback' and a music icon when displaying 'Now Playing'